requirements.txt
python-telegram-bot==22.7
python-dotenv==1.2.2
openai==2.29.0
sentence-transformers==5.3.0
transformers==5.3.0
pillow==12.1.1
torch==2.10.0